+++ wp-video-grid.php
@@ -3,7 +3,7 @@
  * Plugin Name: WordPress Video Grid
  * Plugin URI: https://ajithrn.com/wp-video-grid
  * Description: Plugin to manage and display videos in grid
- * Version: 1.3.0
+ * Version: 1.4.0
  * Author: Ajith
  * Author URI: https://ajithrn.com
  * License: GPL-2.0+
@@ -18,7 +18,7 @@ if ( ! defined( 'WPINC' ) ) {
 }
 
 // Define plugin constants
-define( 'WP_VIDEO_GRID_VERSION', '1.3.0' );
+define( 'WP_VIDEO_GRID_VERSION', '1.4.0' );
 define( 'WP_VIDEO_GRID_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( 'WP_VIDEO_GRID_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 
@@ -31,6 +31,7 @@ require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Admin.php';
 require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Ajax.php';
 require_once WP_VIDEO_GRID_PLUGIN_DIR . 'src/Settings.php';
 require_once WP_VIDEO_GRID_PLUGIN_DIR . 'blocks/video-grid/block.php';
+require_once WP_VIDEO_GRID_PLUGIN_DIR . 'blocks/single-video/block.php';
 
 /**
  * Initialize the plugin
@@ -44,6 +45,7 @@ function wp_video_grid_init() {
     new WP_Video_Grid\Ajax();
     new WP_Video_Grid\Settings();
     new WP_Video_Grid\Blocks\Video_Grid_Block();
+    new WP_Video_Grid\Blocks\Single_Video_Block();
 }
 add_action( 'plugins_loaded', 'wp_video_grid_init' );